+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - go bindings: fix the internal `os.FileMode` to `S_IF*` conversion to not
   auto-include `S_IFREG` for non-`Mknod` operations (previously this would
   cause `MkdirAll` to error out).
+- `Root::create_file` now supports `O_TMPFILE`.
 
 ### Changed ###
 - The `openat2` resolver will now return `-EAGAIN` if the number of `openat2`